packages/datacommons-sample-agents/pyproject.toml
[project]
name = "datacommons-sample-agents"
dynamic = ["version"]
description = "Data Commons Sample Agents"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "google-adk",
    "google-adk[eval]",
    "datacommons-mcp",
]
urls = {Homepage = "https://github.com/datacommonsorg/agent-toolkit"}

[build-system]
requires = ["uv", "setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
datacommons_sample_agents = ["*.json"]

[tool.setuptools.dynamic]
version = {attr = "datacommons_sample_agents.version.__version__"}
